More inspection stuff, new help utils.
Now can view help commands and stuff arranged in their
respective cogs, which is kind of funky I guess, if you
are into boring stuff like that.

Fixed a bunch of pagination bugs, logic errors and stuff
that was generally shit or poorly done. Reacts should be
a little faster now.

Help now actually stores all alias permutations, so you
can look up commands using any permutation of aliases.
Seems I was an idiot and forgot to add this basic
functionality... ~~Sorry.~~

Since we could have a large number of alias permutations,
we now use the magical **cached_property** decorator to
cache this information between reloads of commands. So,
it should be a bit faster, albeit at the cost of some
arbritary quantity of memory that I cannot be bothered to
work out.